Fix compiler errors and warnings (gcc 10)
This involves declaring shared (included) variables as extern in headers
and defining them somewhere. The same for functions. Also define missing
return types on functions. And this required some reordering of
functions inside of the source files so that they appear before they are
used.
No code is changed with this commit. Code / Variables are only moved.